# contracts/providers/finance/finance.openapi.v1.json
{
  "components": {
    "schemas": {
      "CreateFinanceResearchRequest": {
        "additionalProperties": false,
        "properties": {
          "contract_version": {
            "const": 1,
            "title": "Contract Version",
            "type": "integer"
          },
          "constraints": {
            "items": {
              "type": "string"
            },
            "maxItems": 100,
            "title": "Constraints",
            "type": "array"
          },
          "domain_context": {
            "additionalProperties": {
              "$ref": "#/components/schemas/JsonValue"
            },
            "title": "Domain Context",
            "type": "object"
          },
          "horizon_days": {
            "anyOf": [
              {
                "minimum": 1,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Horizon Days"
          },
          "instrument": {
            "maxLength": 20,
            "minLength": 1,
            "title": "Instrument",
            "type": "string"
          },
          "objective": {
            "maxLength": 1000,
            "minLength": 1,
            "title": "Objective",
            "type": "string"
          },
          "quality_metadata": {
            "additionalProperties": {
              "$ref": "#/components/schemas/JsonValue"
            },
            "title": "Quality Metadata",
            "type": "object"
          },
          "request_id": {
            "format": "uuid",
            "title": "Request Id",
            "type": "string"
          },
          "research": {
            "items": {
              "$ref": "#/components/schemas/ResearchMaterialRequest"
            },
            "maxItems": 100,
            "title": "Research",
            "type": "array"
          },
          "source_domain": {
            "maxLength": 64,
            "minLength": 1,
            "title": "Source Domain",
            "type": "string"
          },
          "source_reference": {
            "maxLength": 500,
            "minLength": 1,
            "title": "Source Reference",
            "type": "string"
          },
          "sources": {
            "items": {
              "$ref": "#/components/schemas/SourceMaterialRequest"
            },
            "maxItems": 100,
            "title": "Sources",
            "type": "array"
          },
          "thesis": {
            "anyOf": [
              {
                "maxLength": 1000,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Thesis"
          }
        },
        "required": [
          "contract_version",
          "request_id",
          "source_domain",
          "source_reference",
          "instrument",
          "objective",
          "domain_context"
        ],
        "title": "CreateFinanceResearchRequest",
        "type": "object"
      },
      "FinanceResearchAcceptedResponse": {
        "additionalProperties": false,
        "properties": {
          "confidence": {
            "maximum": 1,
            "minimum": 0,
            "title": "Confidence",
            "type": "number"
          },
          "contract_version": {
            "const": 1,
            "title": "Contract Version",
            "type": "integer"
          },
          "correlation_id": {
            "title": "Correlation Id",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "finance_research_id": {
            "format": "uuid",
            "title": "Finance Research Id",
            "type": "string"
          },
          "issues": {
            "items": {
              "type": "string"
            },
            "title": "Issues",
            "type": "array"
          },
          "instrument": {
            "title": "Instrument",
            "type": "string"
          },
          "recommendation": {
            "title": "Recommendation",
            "type": "string"
          },
          "request_id": {
            "format": "uuid",
            "title": "Request Id",
            "type": "string"
          },
          "replayed": {
            "title": "Replayed",
            "type": "boolean"
          },
          "source_domain": {
            "title": "Source Domain",
            "type": "string"
          },
          "source_reference": {
            "title": "Source Reference",
            "type": "string"
          },
          "status": {
            "const": "accepted",
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "contract_version",
          "finance_research_id",
          "request_id",
          "source_domain",
          "source_reference",
          "instrument",
          "recommendation",
          "confidence",
          "issues",
          "correlation_id",
          "created_at",
          "replayed",
          "status"
        ],
        "title": "FinanceResearchAcceptedResponse",
        "type": "object"
      },
      "JsonScalar": {
        "anyOf": [
          {
            "type": "string"
          },
          {
            "type": "integer"
          },
          {
            "type": "number"
          },
          {
            "type": "boolean"
          },
          {
            "type": "null"
          }
        ]
      },
      "JsonValue": {
        "anyOf": [
          {
            "$ref": "#/components/schemas/JsonScalar"
          },
          {
            "items": {
              "$ref": "#/components/schemas/JsonValue"
            },
            "type": "array"
          },
          {
            "additionalProperties": {
              "$ref": "#/components/schemas/JsonValue"
            },
            "type": "object"
          }
        ]
      },
      "ProblemDetails": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "correlation_id": {
            "title": "Correlation Id",
            "type": "string"
          },
          "detail": {
            "title": "Detail",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "integer"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "type": {
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "detail",
          "code",
          "correlation_id"
        ],
        "title": "ProblemDetails",
        "type": "object"
      },
      "ResearchMaterialRequest": {
        "additionalProperties": false,
        "properties": {
          "facts": {
            "items": {
              "type": "string"
            },
            "maxItems": 100,
            "title": "Facts",
            "type": "array"
          },
          "summary": {
            "maxLength": 2000,
            "minLength": 1,
            "title": "Summary",
            "type": "string"
          }
        },
        "required": [
          "summary"
        ],
        "title": "ResearchMaterialRequest",
        "type": "object"
      },
      "SourceMaterialRequest": {
        "additionalProperties": false,
        "properties": {
          "as_of": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "As Of"
          },
          "source_reference": {
            "maxLength": 500,
            "minLength": 1,
            "title": "Source Reference",
            "type": "string"
          },
          "title": {
            "maxLength": 300,
            "minLength": 1,
            "title": "Title",
            "type": "string"
          },
          "url": {
            "anyOf": [
              {
                "format": "uri",
                "maxLength": 2083,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Url"
          }
        },
        "required": [
          "source_reference",
          "title"
        ],
        "title": "SourceMaterialRequest",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "AI Platform Finance API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/v1/finance-researches": {
      "post": {
        "operationId": "createFinanceResearch",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "maxLength": 200,
              "minLength": 1,
              "title": "Idempotency-Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateFinanceResearchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FinanceResearchAcceptedResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Forbidden"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Create Finance Research",
        "tags": [
          "Finance"
        ]
      }
    }
  }
}